# resources/curation-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Curation REST API",
    "description": "The Curation REST API enables human-in-the-loop review of entity resolution decisions. Curators can browse low-confidence matches, accept or reject proposed canonical entities, assign mentions to alternative clusters, and perform bulk curation actions. The API also provides authentication, user management, audit trails, and registry/curation statistics.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/user-actions": {
      "get": {
        "tags": [
          "User Actions"
        ],
        "summary": "List User Actions",
        "description": "List cursor-paginated user actions with optional filtering.",
        "operationId": "list_user_actions_api_v1_user_actions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserActionType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by type of user action.",
              "title": "Action Type"
            },
            "description": "Filter by type of user action."
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by actor identifier or email.",
              "title": "Actor"
            },
            "description": "Filter by actor identifier or email."
          },
          {
            "name": "time_range_start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include only actions recorded at or after this timestamp.",
              "title": "Time Range Start"
            },
            "description": "Include only actions recorded at or after this timestamp."
          },
          {
            "name": "time_range_end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Include only actions recorded at or before this timestamp.",
              "title": "Time Range End"
            },
            "description": "Include only actions recorded at or before this timestamp."
          },
          {
            "name": "ordering",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BaseOrdering"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort order for the returned actions.",
              "title": "Ordering"
            },
            "description": "Sort order for the returned actions."
          },
          {
            "name": "decision_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by the decision identifier.  Returns only user actions recorded against this decision (matched via the entity mention triad).",
              "title": "Decision Id"
            },
            "description": "Filter by the decision identifier.  Returns only user actions recorded against this decision (matched via the entity mention triad)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor from previous response",
              "title": "Cursor"
            },
            "description": "Pagination cursor from previous response"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paginated list of user actions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_UserActionSummary_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "503": {
            "description": "MongoDB unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-actions/{action_id}/selected-cluster": {
      "get": {
        "tags": [
          "User Actions"
        ],
        "summary": "Get Selected Cluster",
        "description": "Get the selected cluster preview with top entity mentions.",
        "operationId": "get_selected_cluster_api_v1_user_actions__action_id__selected_cluster_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the user action.",
              "title": "Action Id"
            },
            "description": "Unique identifier of the user action."
          }
        ],
        "responses": {
          "200": {
            "description": "The canonical entity cluster selected by the curator, or null if none was selected.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CanonicalEntityPreview"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Selected Cluster Api V1 User Actions  Action Id  Selected Cluster Get"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/user-actions/{action_id}/candidates": {
      "get": {
        "tags": [
          "User Actions"
        ],
        "summary": "Get Candidates",
        "description": "Get paginated candidate cluster previews with top entity mentions.",
        "operationId": "get_candidates_api_v1_user_actions__action_id__candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the user action.",
              "title": "Action Id"
            },
            "description": "Unique identifier of the user action."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated candidate cluster previews for the given user action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResult_CanonicalEntityPreview_"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User",
        "description": "Create a new user (admin only).",
        "operationId": "create_user_api_v1_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Users",
        "description": "List all users (verified users).",
        "operationId": "list_users_api_v1_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Partial email match",
              "title": "Email"
            },
            "description": "Partial email match"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResult_UserResponse_"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Patch User",
        "description": "Update user flags (admin only).",
        "operationId": "patch_user_api_v1_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of the user to update.",
              "title": "User Id"
            },
            "description": "Unique identifier of the user to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseOrdering": {
        "type": "string",
        "enum": [
          "created_at",
          "-created_at"
        ],
        "title": "BaseOrdering",
        "description": "Base ordering options available to all entity listings."
      },
...
      "CanonicalEntityPreview": {
        "properties": {
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id",
            "description": "Unique identifier of the canonical entity cluster."
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score",
            "description": "Model confidence that this cluster is the correct match."
          },
          "similarity_score": {
            "type": "number",
            "title": "Similarity Score",
            "description": "Similarity score between the entity mention and the cluster."
          },
          "cluster_size": {
            "type": "integer",
            "title": "Cluster Size",
            "description": "Total number of decisions (entity mentions) assigned to this cluster, sourced from the cluster_sizes projection."
          },
          "top_entities": {
            "items": {
              "$ref": "#/components/schemas/EntityMentionPreview"
            },
            "type": "array",
            "title": "Top Entities",
            "description": "Representative entity mentions from this cluster."
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "confidence_score",
          "similarity_score",
          "cluster_size",
          "top_entities"
        ],
        "title": "CanonicalEntityPreview",
        "description": "Cluster preview with top entity mentions for display."
      },
...
      "CreateUserRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address for the new user account."
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "Initial plain-text password (8\u2013128 characters); stored hashed."
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the account is active upon creation.",
            "default": true
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "description": "Grant superuser (admin) privileges to the new user.",
            "default": false
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "description": "Mark the user's email as verified upon creation.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "CreateUserRequest",
        "description": "Admin request to create a user."
      },
...
      "CurationErrorResponse": {
        "properties": {
          "error_code": {
            "$ref": "#/components/schemas/CurationErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation of the error."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "ERS business request UUID for log/trace correlation. Populated by handlers that run after the request-id middleware."
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "CurationErrorResponse",
        "description": "Standard error response body returned by all Curation API endpoints.\n\nThe ``request_id`` field carries the ERS business UUID set by the request\nmiddleware (``set_request_id`` in ``ers.commons.adapters.tracing``). It is\npopulated only by handlers that have access to that context \u2014 primarily the\n``Exception`` (HTTP 500) handler \u2014 and is ``None`` on responses produced\nbefore the middleware ran or by handlers that do not need correlation."
      },
...
      "CursorPage_UserActionSummary_": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/UserActionSummary"
            },
            "type": "array",
            "title": "Results",
            "description": "Page of result items."
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of items returned in this page.",
            "default": 0
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque cursor to fetch the next page, or null if there are no more pages."
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "CursorPage[UserActionSummary]"
      },
...
      "PaginatedResult_CanonicalEntityPreview_": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Total number of items matching the query."
          },
          "previous": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous",
            "description": "Previous page number, or null if on the first page."
          },
          "next": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next",
            "description": "Next page number, or null if on the last page."
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CanonicalEntityPreview"
            },
            "type": "array",
            "title": "Results",
            "description": "Page of result items."
          }
        },
        "type": "object",
        "required": [
          "count",
          "results"
        ],
        "title": "PaginatedResult[CanonicalEntityPreview]"
      },
      "PaginatedResult_UserResponse_": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Total number of items matching the query."
          },
          "previous": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous",
            "description": "Previous page number, or null if on the first page."
          },
          "next": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next",
            "description": "Next page number, or null if on the last page."
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "type": "array",
            "title": "Results",
            "description": "Page of result items."
          }
        },
        "type": "object",
        "required": [
          "count",
          "results"
        ],
        "title": "PaginatedResult[UserResponse]"
      },
...
      "UserActionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier of the user action."
          },
          "about_entity_mention": {
            "$ref": "#/components/schemas/EntityMentionPreview"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ClusterReference"
            },
            "type": "array",
            "title": "Candidates",
            "description": "Candidate clusters presented to the curator."
          },
          "selected_cluster": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClusterReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "action_type": {
            "$ref": "#/components/schemas/UserActionType"
          },
          "actor": {
            "$ref": "#/components/schemas/ActorSummary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the user action was recorded."
          },
          "metadata": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Optional additional metadata attached to the action."
          }
        },
        "type": "object",
        "required": [
          "id",
          "about_entity_mention",
          "candidates",
          "action_type",
          "actor",
          "created_at"
        ],
        "title": "UserActionSummary",
        "description": "User action summary for list display."
      },
      "UserActionType": {
        "type": "string",
        "enum": [
          "ACCEPT_TOP",
          "ACCEPT_ALTERNATIVE",
          "REJECT_ALL"
        ],
        "title": "UserActionType",
        "description": "Types of curator actions on entity mention resolutions"
      },
...
      "UserPatchRequest": {
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "Set to true to enable the account or false to disable it."
          },
          "is_superuser": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Superuser",
            "description": "Set to true to grant or false to revoke superuser privileges."
          },
          "is_verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Verified",
            "description": "Set to true to mark the email as verified or false to unverify."
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "New plain-text password (8\u2013128 characters); stored hashed."
          }
        },
        "type": "object",
        "title": "UserPatchRequest",
        "description": "Admin request to update user attributes."
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier of the user."
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Email address of the user."
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the user account is active and can log in."
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "description": "Whether the user has superuser (admin) privileges."
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "description": "Whether the user's email address has been verified."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the user account was created."
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Timestamp of the last update to the user account."
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "is_active",
          "is_superuser",
          "is_verified",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Public user representation (no password)."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}