# src/ers/curation/entrypoints/api/v1/user_actions.py
from datetime import datetime
from typing import Annotated

from erspec.models.core import UserActionType
from fastapi import APIRouter, Depends, Path, Query

from ers.commons.domain.data_transfer_objects import CursorPage, PaginatedResult
from ers.curation.domain.data_transfer_objects import (
    BaseOrdering,
    CanonicalEntityPreview,
    UserActionFilters,
    UserActionSummary,
)
from ers.curation.entrypoints.api.auth import VerifiedUser
from ers.curation.entrypoints.api.dependencies import (
    get_canonical_entity_service,
    get_user_action_service,
)
from ers.curation.entrypoints.api.v1.schemas import CursorPagination, ErrorResponse, Pagination
from ers.curation.services import CanonicalEntityService, UserActionService

router = APIRouter(prefix="/user-actions", tags=["User Actions"])


@router.get(
    "",
    responses={
        400: {"model": ErrorResponse},
        403: {"model": ErrorResponse},
        503: {"model": ErrorResponse, "description": "MongoDB unavailable"},
    },
    response_description="Cursor-paginated list of user actions.",
)
async def list_user_actions(
    cursor_params: CursorPagination,
    _user: VerifiedUser,
    service: Annotated[UserActionService, Depends(get_user_action_service)],
    action_type: Annotated[
        UserActionType | None, Query(description="Filter by type of user action.")
    ] = None,
    actor: Annotated[str | None, Query(description="Filter by actor identifier or email.")] = None,
    time_range_start: Annotated[
        datetime | None,
        Query(description="Include only actions recorded at or after this timestamp."),
    ] = None,
    time_range_end: Annotated[
        datetime | None,
        Query(description="Include only actions recorded at or before this timestamp."),
    ] = None,
    ordering: Annotated[
        BaseOrdering | None, Query(description="Sort order for the returned actions.")
    ] = None,
    decision_id: Annotated[
        str | None,
        Query(
            description=(
                "Filter by the decision identifier.  Returns only user actions recorded "
                "against this decision (matched via the entity mention triad)."
            )
        ),
    ] = None,
) -> CursorPage[UserActionSummary]:
    """List cursor-paginated user actions with optional filtering."""
    filters = None
    if any(
        v is not None
        for v in (action_type, actor, time_range_start, time_range_end, ordering, decision_id)
    ):
        filters = UserActionFilters(
            action_type=action_type,
            actor=actor,
            time_range_start=time_range_start,
            time_range_end=time_range_end,
            ordering=ordering,
            decision_id=decision_id,
        )
    return await service.list_user_actions(cursor_params, filters)


@router.get(
    "/{action_id}/selected-cluster",
    responses={
        403: {"model": ErrorResponse},
        404: {"model": ErrorResponse},
    },
    response_description="The canonical entity cluster selected by the curator, or null if none was selected.",
)
async def get_selected_cluster(
    action_id: Annotated[str, Path(description="Unique identifier of the user action.")],
    _user: VerifiedUser,
    service: Annotated[UserActionService, Depends(get_user_action_service)],
    canonical_service: Annotated[CanonicalEntityService, Depends(get_canonical_entity_service)],
) -> CanonicalEntityPreview | None:
    """Get the selected cluster preview with top entity mentions."""
    return await service.get_selected_cluster_preview(action_id, canonical_service)


@router.get(
    "/{action_id}/candidates",
    responses={
        403: {"model": ErrorResponse},
        404: {"model": ErrorResponse},
    },
    response_description="Paginated candidate cluster previews for the given user action.",
)
async def get_candidates(
    action_id: Annotated[str, Path(description="Unique identifier of the user action.")],
    pagination: Pagination,
    _user: VerifiedUser,
    service: Annotated[UserActionService, Depends(get_user_action_service)],
    canonical_service: Annotated[CanonicalEntityService, Depends(get_canonical_entity_service)],
) -> PaginatedResult[CanonicalEntityPreview]:
    """Get paginated candidate cluster previews with top entity mentions."""
    return await service.get_candidate_previews(action_id, pagination, canonical_service)